requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.2
cssselect==1.2.0
exceptiongroup==1.1.2
filelock==3.12.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.12.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
multidict==6.0.4
openai==0.27.8
packaging==23.1
parsel==1.8.1
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.9.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tldextract==3.4.4
tornado==6.3.2
tqdm==4.65.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
w3lib==2.1.1
yarl==1.9.2
zope.interface==6.0